requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.1.0
boto3==1.39.4
botocore==1.39.4
cachetools==5.5.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.2.1
colorama==0.4.6
comm==0.1.4
contourpy==1.3.1
cryptography==42.0.2
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.4
django-storages==1.14.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.0.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fastjsonschema==2.18.1
fonttools==4.55.6
fqdn==1.5.1
frozenlist==1.4.1
google-api-core==2.24.0
google-api-python-client==2.160.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.2.3
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.4
importlib-metadata==7.0.1
iniconfig==2.1.0
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jaraco.classes==3.3.1
jedi==0.19.1
Jinja2==3.1.6
jira==3.8.0
jmespath==1.0.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keyring==24.3.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
multidict==6.0.5
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
oauthlib==3.2.2
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
pillow==11.1.0
platformdirs==3.11.0
pluggy==1.6.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.25.0
protobuf==5.29.2
psutil==5.9.6
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.16.1
PyMySQL==1.1.1
pyparsing==3.2.1
pytest==8.4.1
pytest-asyncio==1.0.0
python-dateutil==2.8.2
python-dotenv==1.1.1
python-jose==3.5.0
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
pyxlsb==1.0.10
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rpds-py==0.10.6
rsa==4.9
s3transfer==0.13.0
Send2Trash==1.8.2
sentry-sdk==2.33.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.1
terminado==0.17.1
tinycss2==1.2.1
tools==1.0.2
tornado==6.3.3
traitlets==5.11.2
typer==0.16.0
types-python-dateutil==2.8.19.14
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==15.0.1
widgetsnbextension==4.0.9
XlsxWriter==3.1.9
yarl==1.9.4
zipp==3.17.0
